Move detubmle into GNC
