+++ state_machine/applications/flight/StateMachineConfig.py
@@ -6,7 +6,7 @@ from Tasks.time_task import task as time
 from Tasks.test_task import task as test
 from Tasks.lowpower5 import task as lowpower5
 from Tasks.lowpower5later import task as lowpower5later
-from Tasks.detumble import task as detumble
+from Tasks.gnc import task as gnc
 
 from TransitionFunctions import announcer, low_power_on, low_power_off
 
@@ -19,7 +19,7 @@ TaskMap = {
     "Test": test,
     "LowPower5": lowpower5,
     "LowPower5Later": lowpower5later,
-    'Detumble': detumble,
+    "GNC": gnc,
 }
 
 TransitionFunctionMap = {
@@ -61,7 +61,7 @@ config = {
                 'Priority': 1,
                 'ScheduleLater': False
             },
-            'Detumble': {
+            'GNC': {
                 'Interval': 0.1,
                 'Priority': 3,
                 'ScheduleLater': False
