--- state_machine/applications/flight/Tasks/mekf.py
@@ -1,10 +1,0 @@
-from lib.template_task import Task
-
-
-class task(Task):
-    """Multiplicative Extended Kalman Filter"""
-    name = 'mekf'
-    color = 'red'
-
-    async def main_task(self):
-        pass
