+++ state_machine/unit_tests/test_control.py
@@ -5,10 +5,10 @@ from numpy import testing
 sys.path.insert(0, './state_machine/applications/flight')
 sys.path.insert(0, './state_machine/frame')
 
-from Tasks.detumble import bcross  # noqa: E402
+from lib.control import bcross  # noqa: E402
 
 
-class MiscTests(unittest.TestCase):
+class BCrossTest(unittest.TestCase):
 
     def test(self):
         b = [0, 2, 3]
